Refactor frontend into modular ES scripts
