+++ public/js/state.js
@@ -1,0 +1,8 @@
+export const state = {
+  pageCounter: 0,
+  saveTimeout: null,
+  isUpdatingFromServer: false,
+  saveIndicator: null,
+  lastSyncedSignature: null,
+  pageStreamSource: null,
+};
